#246 feat: Update sidebar design
